LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/row-auto-repeat-auto-020-expected.txt


FAIL Auto-fit repeat tracks with auto size and auto item placement assert_equals: expected "100px 100px 100px 100px 100px 500px 0px 0px 0px 0px" but got "0px 0px 0px 0px 0px 600px"
